restructured folder organization for components
